Create exercise12.py
A toy store is very successful in two of its products: clowns and dolls. They usually sell by mail order and the logistics company charges them by weight for each package so they have to calculate the weight of the clowns and dolls that will be in each package on demand. Each clown weighs 112 gr and each doll 75 gr. Write a program that reads the number of clowns and dolls sold in the last order and calculates the total weight of the package that will be sent.