Create exercise13.py
Imagine that you have just opened a new savings account that offers 4% interest per year. These savings due to interest, which is not charged until the end of the year, are added to your final savings account balance. Write a program that starts by reading the amount of money deposited in the savings account, entered by the user. Then the program should calculate and display the amount of savings after the first, second and third years. Round each amount to two decimal places.