Create exercise14.py
A bakery sells loaves of bread at 3.49 euros each. 
Bread that is not the day has a 60% discount. 
Write a program that starts by reading the number of bars sold that are not of the day. 
Then the program should show the usual price of a loaf of bread, 
the discount you get for not being fresh and the total final cost.